\name{fpca_res}
\alias{fpca_res}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculation of losses of M and L for FPCA
}
\description{
Compare the original data and its reconstruction with calculation of losses of M and L for FPCA
}
\usage{
fpca_res(tnsr, smooth_tnsr, true_L, npc = NULL, pve = 0.99, center = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{tnsr}{
  A tensor of dimension a*b*n. Should be organized so that the first two dimensions are to be compressed.
}
  \item{smooth_tnsr}{
  A tensor of dimension a*b*n, which contains the smooth, complete ground-truth data.
}
  \item{true_L}{
  Ground-truth L matrix.
}
  \item{npc}{
  Number of principal components for FPCA. Can be NULL (flexible ranks) or a certain number. 
}
  \item{pve}{
  The percent variance explained by the returned number of PCs. If not defined, the default is 0.99. 
}
  \item{center}{
  Logical: determine whether to center the data or not.
}
}
\details{
This function is developed based on the fpca.sc() function in the refund package (Goldsmith et al., 2024) and the rTensor package (Li et al., 2018). 
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
\item{est}{An a*b*n array of estimation generated by FPCA}
\item{Lfpca}{L matrix generated by FPCA}
\item{loss_M}{Losses in the reconstruction of M}
\item{loss_L}{Losses in the reconstruction of L}
%% ...
}
\references{
Goldsmith J, Scheipl F, Huang L, Wrobel J, Di C, Gellar J, Harezlak J, McLean MW, Swihart B,
Xiao L, Crainiceanu C, Reiss PT, Cui E (2024).
\emph{refund: Regression with Functional Data}.
R package version 0.1-37. \url{https://CRAN.R-project.org/package=refund}.
\doi{10.32614/CRAN.package.refund}

Li J, Bien J, Wells MT (2018).
\emph{rTensor: An R Package for Multidimensional Array (Tensor) Unfolding, Multiplication, and Decomposition}.
\emph{Journal of Statistical Software}, \bold{87}(10), 1--31. \doi{10.18637/jss.v087.i10}
}
\author{
Leyuan Qian
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
%\examples{

%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
