\name{SmoothHOOI-package}
\alias{SmoothHOOI-package}
\alias{SmoothHOOI}
\docType{package}
\title{\packageTitle{SmoothHOOI}}
\description{\packageDescription{SmoothHOOI}}
\section{Package Content}{\packageIndices{SmoothHOOI}}
\author{\packageAuthor{SmoothHOOI}}
\section{Maintainer}{\packageMaintainer{SmoothHOOI}}
\keyword{package}
